declare one class per file
